Use AUR package swaylock-effects
There is probably a solution for non-nixos systems to use the nix package for swaylock-effects, but it would require changing mount options for /run/wrappers. I am not comfortable with doing that and the AUR package is fine and has worked for ages.
See alternative solution here: https://github.com/NixOS/nixpkgs/issues/158025#issuecomment-1616807870